+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Dynamic theme for Turnstile widget. 
+
 ## 0.1.0 - 2024-10-16
 
 - Initial public release.